Downgrade javafx to 20.0.1 (#246)
* Downgraded openjfx.version in pom.xml

The openjfx.version in the pom.xml file has been downgraded from 21.0.1 to 20.0.1. This change is made to address UI test failures on unix systems with the previous version, thus improving the reliability and compatibility of the project.

* Downgrade openjfx.version in pom.xml

The openjfx.version in pom.xml was reverted to 20.0.1 to resolve UI test failures on Unix systems. The old version negatively impacted reliability and compatibility. The downgrade should enhance the project's overall stability.